+++ lib/src/main/kotlin/com/kimdb/KImdb.kt
@@ -4,6 +4,7 @@ import com.kimdb.api.KImdbRepository
 import com.kimdb.model.Name
 import com.kimdb.model.Title
 import com.kimdb.tsv.ImdbTsvParser
+import java.io.InputStream
 import java.nio.file.Path
 
 object KImdb {
@@ -16,6 +17,11 @@ object KImdb {
         titleBasicsPath: Path,
         nameBasicsPath: Path
     ): KImdbRepository = ImdbTsvParser.loadRepository(titleBasicsPath, nameBasicsPath)
+
+    fun inMemoryRepositoryFromTsv(
+        titleBasicsStream: InputStream,
+        nameBasicsStream: InputStream
+    ): KImdbRepository = ImdbTsvParser.loadRepository(titleBasicsStream, nameBasicsStream)
 
     fun sqliteRepository(
         dbPath: Path